+++ CHANGES.md
@@ -5,6 +5,8 @@
         - onply search on positive strand when searching unlimited or protein sequences
     - `seqkit locate`:
         - removing debug info for `-r` introduced in a0f6b6e. [#180](https://github.com/shenwei356/seqkit/issues/180)
+    - `seqkit fx2tab`:
+        - new flag `-C/--base-count` for counting bases.
 - [SeqKit v0.15.0](https://github.com/shenwei356/seqkit/releases/tag/v0.15.0)
 [![Github Releases (by Release)](https://img.shields.io/github/downloads/shenwei356/seqkit/v0.15.0/total.svg)](https://github.com/shenwei356/seqkit/releases/tag/v0.15.0)
     - `seqkit grep/locate`: update help message.